Adds scenes and acts to avoid update function override in main
